lingo-stage1-audio-only/lingo-v2/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
livekit-api==0.7.1
pydantic==2.10.1
